+++ gitspy/__init__.py
@@ -1,6 +1,5 @@
 """Intuitive Git for Python."""
-from . import exceptions
 from ._subprocess import Git
 from ._version import __version__
 
-__all__ = ["__version__", "exceptions", "Git"]
+__all__ = ["__version__", "Git"]