+++ gitspy/_subprocess.py
@@ -9,7 +9,6 @@ from pathlib import Path as _Path
 
 from spall import Subprocess as _Subprocess
 
-from . import exceptions as _exceptions
 from ._environ import TempEnvVar as _TempEnvVar
 
 
@@ -29,6 +28,19 @@ class Git(_Subprocess):
         ]
         super().__init__("git", positionals=self.commands)
 
+    def _get_gitdir(self, path: _Path) -> _t.Optional[_Path]:
+        # find and return the path to the repository's .git directory
+        # start with provided path and any parent up to / where None is
+        # returned if working outside a repository
+        gitdir = path / ".git"
+        if gitdir.is_dir():
+            return gitdir
+
+        if path == _Path("/"):
+            return None
+
+        return self._get_gitdir(path.parent)
+
     def call(self, *args: str, **kwargs: _t.Union[bool, str]) -> int:
         """Call partial git command instantiated in superclass.
 
@@ -37,21 +49,15 @@ class Git(_Subprocess):
         :key capture: Pipe stream to self.
         :key devnull: Suppress output.
         :key suppress: Suppress errors and continue running.
-        :raises NotARepositoryError: If not run from within a
-            repository.
         :raises CalledProcessError: If error occurs in subprocess.
         :return: Exit status.
         """
-        git_dir = _Path.cwd() / ".git"
-        with _TempEnvVar(GIT_WORK_TREE=str(_Path.cwd()), GIT_DIR=str(git_dir)):
+        cwd = _Path.cwd().absolute()
+        gitdir = self._get_gitdir(cwd) or cwd / ".git"
+        with _TempEnvVar(
+            GIT_WORK_TREE=str(gitdir.parent), GIT_DIR=str(gitdir)
+        ):
             if "--bare" in args:
                 del _os.environ["GIT_WORK_TREE"]
 
-            try:
-                return super().call(*args, **kwargs)
-
-            except _sp.CalledProcessError as err:
-                if not git_dir.is_dir():
-                    raise _exceptions.NotARepositoryError from err
-
-                raise err
+            return super().call(*args, **kwargs)